fix: Correct test DTO field names and method signatures
- Fix AdminDashboardControllerTest: avgDeliveryTimeToday, overallStatus,
  apiLatencyP50, dbLoad, individual order count fields
- Fix AdminDashboardIntegrationTest: Integer types for latency fields
- Fix CustomerExperienceAnalytics tests: ratingCount -> reviewCount
- Fix FraudAnalytics tests: LoginSecurityMetricsDto, RefundAbuseMetricsDto
- Fix TechnicalAnalyticsIntegrationTest: Long type for payloadSizeBytes

+++ src/test/java/com/fooddelivery/analytics/technical/integration/TechnicalAnalyticsIntegrationTest.java
@@ -375,7 +375,7 @@ class TechnicalAnalyticsIntegrationTest {
                     .messageId(UUID.randomUUID().toString())
                     .sessionId("session_" + (i % 50))
                     .messageType(i % 3 == 0 ? "LOCATION_UPDATE" : (i % 3 == 1 ? "ORDER_STATUS" : "CHAT"))
-                    .payloadSizeBytes(100 + (int) (Math.random() * 900))
+                    .payloadSizeBytes((long) (100 + (int) (Math.random() * 900)))
                     .publishedAt(testStart.plusMinutes(i % 60))
                     .deliveredAt(i < 195 ? testStart.plusMinutes(i % 60).plusNanos((long) (Math.random() * 100_000_000)) : null)
                     .isDelivered(i < 195)
